Extract DataFileNamespace to top-level class
This slightly reduces what's going on in Alces, making it marginally
clearer.
